# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Oraclous Platform API
  version: 1.0.0
  description: >-
    The public REST contract served by the application-gateway-service — the platform's single
    external edge (ADR-015, ADR-018). Every request is authenticated once at the edge (a bearer JWT
    issued by the identity service); the gateway forwards a verified identity downstream and streams the
    upstream response back. Every 4xx/5xx the caller sees is the canonical **error envelope**
    (`components/schemas/ErrorEnvelope`) — gateway-own errors and normalised upstream errors alike;
    an upstream body is never relayed, so no stack/SQL/internal-host detail can leak (Interface
    Contracts §3).


    **Stability (ADR-015 §2):** each operation carries `x-stability: stable | provisional`. A `stable`
    operation's shape is locked; breaking changes require a new major version. The `openapi-diff-gate`
    CI job blocks breaking changes to `stable` operations.


    **Evolving contract (R6):** this is the R6-in-progress public surface and grows per release slice
    (integration keys + published agents, chat, webhooks, MCP land in later R6 slices and are added
    here as they ship). The published `openapi/v1.yaml` — not any planning table — is the source of
    truth for what the gateway actually exposes.


    **As-built path note:** the proxied paths are currently heterogeneous (`/v1/...`, `/api/v1/...`,
...
    actually route today.


...


...
servers:
  - url: /
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/graph/{graphId}/temporal:
    get:
      tags: [knowledge-retriever]
      summary: Point-in-time entities whose validity window covers `as_of`.
      operationId: temporalRead
      x-stability: stable
      parameters:
        - { name: graphId, in: path, required: true, schema: { type: string } }
        - { name: as_of, in: query, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /v1/graph/{graphId}/similar/{nodeId}:
    get:
      tags: [knowledge-retriever]
      summary: Entities similar to a node — its SIMILAR_TO neighbours ranked by cosine (#310).
      description: >-
        Traverses the SIMILAR_TO edges the KGS similarity pass wrote and returns the neighbours of
        `nodeId`, ranked by the stamped cosine, as the uniform NodeResult envelope (the cosine +
        relationship live in `properties`). `min_score` 0 returns every link; raise it to keep only
        close ones. Backs the registry `core/find-similar` in-loop tool.
      operationId: findSimilar
      x-stability: stable
      parameters:
        - { name: graphId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: nodeId, in: path, required: true, schema: { type: string } }
        - { name: top_k, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 10 } }
        - { name: min_score, in: query, required: false, schema: { type: number, minimum: 0, maximum: 1, default: 0 } }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
...
  /api/v1/tools:
    get:
      tags: [capabilities]
      summary: The capability catalogue (tools) visible to the caller's organisation.
      operationId: listTools
      x-stability: stable
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
    post:
      tags: [capabilities]
      summary: Register a tool descriptor (idempotent — the id is derived from the descriptor identity).
      operationId: registerTool
      x-stability: stable
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        "201":
          description: The registered tool descriptor.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DescriptorResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/tools/{toolId}:
    get:
      tags: [capabilities]
      summary: Read one tool descriptor (org-scoped; an unknown / cross-org id is a 404 — mask).
      operationId: getTool
      x-stability: stable
      parameters:
        - { name: toolId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The tool descriptor.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DescriptorResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/tools/import-mcp:
    post:
      tags: [capabilities]
      summary: >-
        Import an external MCP server's tools (ADMIN) — each is registered `pending_approval` and is
        not executable until approved. The server URL is SSRF-egress-checked.
      operationId: importMcpTools
      x-stability: stable
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        "201":
          description: The imported `pending_approval` descriptors.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JsonObject" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "5XX": { $ref: "#/components/responses/Upstream5xx" }
  /api/v1/tools/{toolId}/approve:
    post:
      tags: [capabilities]
      summary: Approve an imported MCP tool (ADMIN) — pending_approval → active, making it executable.
      operationId: approveTool
      x-stability: stable
      parameters:
        - { name: toolId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Approved — the tool is now active. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/executions/{executionId}:
    get:
      tags: [capabilities]
      summary: Read an execution row (status + result + provenance), org-scoped.
      operationId: getExecution
      x-stability: stable
      parameters:
        - { name: executionId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/executions:
    post:
      tags: [capabilities]
      summary: Execute a tool synchronously (sync only; durable/async is the engine).
      operationId: executeTool
      x-stability: stable
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /v1/harnesses/execute:
    post:
      tags: [harness]
      summary: Run an OHM to completion or escalation, synchronously; returns the execution row.
      operationId: harnessExecute
      x-stability: stable
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        "201": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "5XX": { $ref: "#/components/responses/Upstream5xx" }
  /v1/harnesses/executions:
    get:
      tags: [harness]
      summary: List the organisation's harness executions.
      operationId: listExecutions
      x-stability: stable
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /v1/harnesses/spend:
    get:
      tags: [harness]
      summary: An ESTIMATE of the org's BYOM provider LLM spend — NOT a platform charge.
      description: >-
        An **estimate** of the caller organisation's own provider (BYOM) LLM spend, priced at read
        time from a static rate table — this is **not** platform billing or a charge Oraclous levies
        (ADR-009: the substrate meters raw tokens; pricing is a downstream read-time concern). Raw
        per-model token sums are priced when the table knows the model; a model **absent** from the
        table reports its tokens only (`estimated_usd: null`, `priced: false`) and is listed in
        `unpriced_models`. `total_estimated_usd` sums only the priced rows. Org-scoped.
      operationId: harnessSpend
      x-stability: provisional
      parameters:
        - name: since
          in: query
          required: false
          description: ISO8601 timestamp; bounds the spend window to executions at or after it. Omit for all-time.
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: The org's estimated BYOM spend, per model and in total.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SpendResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/harnesses/{executionId}/resume:
    post:
      tags: [harness]
      summary: Resolve a mid-loop HITL pause — APPROVED resumes the loop, DENIED fails it.
      operationId: harnessResume
      x-stability: stable
      parameters:
        - { name: executionId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/UpstreamJson" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        A short-lived access token minted by the identity service (member credentials). The gateway verifies
        it once at the edge and forwards a verified identity downstream (ADR-018). Integration-key and
        agent-credential auth modes are added in later R6 slices.
...
  schemas:
    JsonObject:
      type: object
      description: A free-form JSON object; the precise request shape is owned by the upstream service.
      additionalProperties: true
...
    DescriptorResponse:
      type: object
      description: >-
        A capability/tool descriptor as returned by the registry. The full body is owned by the
        capability-registry-service; only the load-bearing fields are named here. `status`
        distinguishes an executable descriptor from one awaiting the MCP-import supply-chain gate.
      additionalProperties: true
      properties:
        id: { type: string, format: uuid }
        organisation_id: { type: string, format: uuid }
        kind: { type: string, description: The descriptor kind (e.g. tool). }
        name: { type: [string, "null"] }
        descriptor:
          type: object
          additionalProperties: true
          description: The OHM manifest JSONB.
        status:
          type: string
          enum: [active, pending_approval]
          default: active
          description: >-
            `active` is executable; `pending_approval` is an imported MCP tool awaiting an admin's
            approval (it is not executable until approved).
...
    ModelSpendOut:
      type: object
      additionalProperties: false
      description: >-
        Per-model spend estimate. `estimated_usd` is null and `priced` is false for a model absent
        from the static rate table — it reports its raw tokens only, never a fabricated price.
      required: [model, input_tokens, output_tokens, executions, estimated_usd, priced]
      properties:
        model:
          type: [string, "null"]
          description: The provider model id these tokens were billed against (null when unattributed).
        input_tokens: { type: integer, description: Summed prompt tokens for this model over the window. }
        output_tokens: { type: integer, description: Summed completion tokens for this model over the window. }
        executions: { type: integer, description: Number of harness executions that used this model. }
        estimated_usd:
          type: [number, "null"]
          description: Estimated cost of this model's tokens; null when the model is not in the rate table.
        priced:
          type: boolean
          description: True when the rate table priced this model; false when only raw tokens are reported.
    SpendResponse:
      type: object
      additionalProperties: false
      description: >-
        An ESTIMATE of the org's provider (BYOM) LLM spend, priced from a static rate table — NOT
        platform billing. Org-scoped. Unpriced models (absent from the table) contribute raw tokens
        only and are named in `unpriced_models`; `total_estimated_usd` sums only the priced rows.
        Mirrors harness-runtime-service `SpendResponse`/`ModelSpendOut`.
      required:
        - since
        - currency
        - by_model
        - total_estimated_usd
        - total_input_tokens
        - total_output_tokens
        - unpriced_models
      properties:
        since:
          type: [string, "null"]
          format: date-time
          description: The window lower bound echoed from the `since` query param; null for all-time.
        currency:
          type: string
          default: USD
          description: The currency the estimate is denominated in (always USD today).
        by_model:
          type: array
          description: One row per model that recorded tokens over the window.
          items: { $ref: "#/components/schemas/ModelSpendOut" }
        total_estimated_usd:
          type: number
          description: Sum of `estimated_usd` over the PRICED rows only (unpriced models contribute zero).
        total_input_tokens: { type: integer, description: Summed prompt tokens across all models. }
        total_output_tokens: { type: integer, description: Summed completion tokens across all models. }
        unpriced_models:
          type: array
          description: Model ids that recorded tokens but were absent from the rate table (tokens-only).
          items: { type: string }
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [field, issue]
      properties:
        field:
          type: string
          minLength: 1
          description: The offending field name or path. Never the value.
        issue:
          type: string
          pattern: "^[A-Z][A-Z0-9_]*$"
          description: Machine-readable issue token (e.g. INVALID_FORMAT). Describes the problem, not the value.
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      description: >-
        Canonical error envelope for every 4xx/5xx the gateway produces (own + normalised upstream).
        `additionalProperties` is false at every level so no stack/cause/exception/trace field can ever
        exist. Mirrors packages/errors/contract/error-envelope.schema.json (the cross-repo source).
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId, retryable]
          properties:
            code:
              type: string
              description: Closed machine-readable taxonomy; the only branch key. Never branch on message.
              enum:
                - VALIDATION_FAILED
                - MALFORMED_REQUEST
                - UNAUTHENTICATED
                - UNAUTHORIZED
                - NOT_FOUND
                - METHOD_NOT_ALLOWED
                - CONFLICT
                - CREDENTIALS_REQUIRED
                - PAYLOAD_TOO_LARGE
                - UNSUPPORTED_MEDIA_TYPE
                - RATE_LIMITED
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
                - GATEWAY_TIMEOUT
            message:
              type: string
              minLength: 1
              description: Curated, generic, human-readable. Never an exception message; never reflects request content.
            requestId:
              type: string
              pattern: "^req_[0-9A-Za-z]+$"
              description: Opaque correlation handle; the only trace exposed. Maps server-side to the full trace.
            retryable:
              type: boolean
              description: Server-authoritative — may a retry succeed? The FE reads this boolean; it never hardcodes which codes are retryable.
            details:
              type: array
              minItems: 1
              description: Present only for VALIDATION_FAILED. Closed sub-vocabulary; never the raw offending value.
              items: { $ref: "#/components/schemas/ErrorDetail" }
            needs_credential:
              type: object
              additionalProperties: false
              required: [requirement_id, provider]
              description: Present only for CREDENTIALS_REQUIRED. Names the missing credential requirement + provider so the FE can prompt onboarding. Leak-safe machine tokens — never a value, URL, host, or secret.
              properties:
                requirement_id:
                  type: string
                  minLength: 1
                  maxLength: 64
                  pattern: "^[A-Za-z0-9][A-Za-z0-9_.-]*$"
                  description: The missing requirement's type token, e.g. api_key. Never a value.
                provider:
                  type: string
                  minLength: 1
                  maxLength: 48
                  pattern: "^[A-Za-z0-9][A-Za-z0-9_.-]*$"
                  description: The provider the credential is for, e.g. web_search. Never a value.
  responses:
    UpstreamJson:
      description: Success — a JSON body produced and owned by the upstream service, streamed through unchanged.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/JsonObject" }
    Unauthenticated:
      description: No or invalid credentials (UNAUTHENTICATED).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Authenticated but not permitted (UNAUTHORIZED).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: No such resource, or an unrouted path (NOT_FOUND).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: The request conflicts with current state (CONFLICT).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ValidationFailed:
      description: The request failed validation (VALIDATION_FAILED; carries details[]).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    Upstream5xx:
      description: An upstream failure normalised to the envelope (SERVICE_UNAVAILABLE / GATEWAY_TIMEOUT / INTERNAL_ERROR).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...